fix: refactor 'Record' to 'Patient' and update file structure
Renamed the 'Record' model, API service, and related files to 'Patient' for improved clarity and consistency. Updated imports, method names, and references throughout the codebase. Moved record-list component files out of the data-table directory to a dedicated record-list directory.
